backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
sqlalchemy
openai
python-dotenv
alembic
psycopg2-binary
passlib
python-jose[cryptography]